+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 > The format is based on [Keep a Changelog](https://keepachangelog.com/de/1.0.0/).
 
 ---
-## [2.0.1] - 2026-01-08
+## [2.0.1] - 2026-01-12
 ### 🔄 Changed
 - License Check
   - Treat `no_license` antivirus state as `OK` during inventory
+- HA Peer check refactored
+  - only one service will be created (run service discovery!)
+  - if no secondary information is found, state will become `WARN`
 - Readme updated
 
 ### 🚀 Added